added point factory
