+++ src/graph/graph.py
@@ -218,87 +218,6 @@ class Graph:
     def get_neighbour_vector(self, label) -> np.array:
         return np.unpackbits(self.adj_neighbour[label])
 
-    ### Generating points on graph ###
-
-    def gen_random_points(self) -> None:
-        self.points = np.array(
-            [self._create_random_point() for _ in range(self.number_of_points)]
-        )
-        pass
-
-    def gen_iko_points(self, divisions=4):
-        iko = ikosaeder()
-        iko.subdivide(n=divisions)
-        self.points = iko.normalized_points()
-        self.number_of_points = len(self.points)
-        pass
-
-    def gen_archimedic_spiral(
-        self,
-        speed=100,
-        N=1000,
-        lower_bound=0,
-        upper_bound=1,
-    ):
-        """theta in [arccos(lower_bound), arccos(upperbound)]"""
-
-        X = np.linspace(lower_bound, upper_bound, N)
-        thetas = np.arccos(X)
-        phis = speed * thetas
-
-        arr = np.array(
-            [
-                np.ones(len(thetas)),
-                thetas,
-                phis,
-                np.sin(thetas) * np.cos(phis),
-                np.sin(thetas) * np.sin(phis),
-                np.cos(thetas),
-            ]
-        )
-        # np.transpose wechselt die Dimmensionen, sodass bei Iterationen über die Zeilen(einzelne Punkte) und nicht die Spalten(Theta-vektor, Phi-vektor, ...) iteriert wird
-        self.points = np.transpose(arr)
-        pass
-
-    # sieht für mich nach der archimedischen Spirale aus mit speed L = sqrt(N*pi)
-    def gen_bauer_spiral(self, N):
-        L = np.sqrt(N * np.pi)
-        k = np.array(range(1, N + 1))
-        z = 1 - (2 * k - 1) / N
-        phi = np.arccos(z)
-        theta = L * phi
-        x = np.sin(phi) * np.cos(theta)
-        y = np.sin(phi) * np.sin(theta)
-
-        arr = np.array(
-            [
-                np.ones(N),
-                phi,
-                theta,
-                np.sin(phi) * np.cos(theta),
-                np.sin(phi) * np.sin(theta),
-                z,
-            ]
-        )
-
-        # np.transpose wechselt die Dimmensionen, sodass bei Iterationen über die Zeilen(einzelne Punkte) und nicht die Spalten(Theta-vektor, Phi-vektor, ...) iteriert wird
-        self.points = np.transpose(arr)
-        pass
-
-    def _create_point(self, r, theta, phi) -> np.array:
-        arr = np.array([r, theta, phi, 0, 0, 0])
-        arr[3] = r * math.sin(theta) * math.cos(phi)
-        arr[4] = r * math.sin(theta) * math.sin(phi)
-        arr[5] = r * math.cos(theta)
-        return arr
-
-    def _create_random_point(self) -> np.array:
-        phi = random() * 2 * math.pi
-        x = random() * 2 - 1
-        theta = np.arccos(x)
-
-        return self._create_point(1, theta, phi)
-
 
 pass
